Allow values for limitInputPixels larger than 32-bit #3238

+++ src/common.h
@@ -49,7 +49,7 @@ namespace sharp {
     std::string file;
     char *buffer;
     VipsFailOn failOn;
-    int limitInputPixels;
+    uint64_t limitInputPixels;
     bool unlimited;
     VipsAccess access;
     size_t bufferLength;
